refactor: Replace old CLI with streaming version
- Remove legacy cli.ts
- Rename cli-stream.ts to cli.ts
- All transcriptions now use streaming by default
- Google Drive URL support included in main CLI
